% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatBankData.R
\name{getStatBankData}
\alias{getStatBankData}
\title{Retrieve data from CSO's PxStat Open Data Platform database.}
\usage{
getStatBankData(table, metadata = TRUE, type = "px")
}
\arguments{
\item{table}{A string}

\item{metadata}{A boolean value}

\item{type}{A string}
}
\value{
px, data.frame, ts or mts
}
\description{
Retrieve data from CSO's PxStat Open Data Platform database.
}
