% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px2col.R
\name{px2col}
\alias{px2col}
\title{Convert px Object to Column-Orientated Dataframe}
\usage{
px2col(px)
}
\arguments{
\item{px}{A px object (from pxR package)}
}
\value{
A data.frame
}
\description{
Convert px Object to Column-Orientated Dataframe
}
