% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px2ts.R
\name{px2ts}
\alias{px2ts}
\title{Convert px Object to ts (or mts) Object}
\usage{
px2ts(px)
}
\arguments{
\item{px}{A px object (from pxR package)}
}
\value{
A ts (or mts) object
}
\description{
Convert px Object to ts (or mts) Object
}
